tools/hotplug: don't install init.d scripts if SYSTEMD
When Xen is configured with --with-systemd currently both the systemd
init scripts as well as the traditional init.d scripts (e.g. xencommons)
are installed.

This causes issues to distros where old style init scripts are still
supported even when systemd is enabled, e.g. Yocto. The consequence is
that xenconsoled (and other daemons) are executed twice resulting in
errors at boot.

If --with-systemd don't install xencommons and the other sysvinit
scripts.

Signed-off-by: Stefano Stabellini <stefano.stabellini@xilinx.com>
Acked-by: Chris Kohn <ckohn@xilinx.com>