xen/eemi: on CLOCK_ENABLE, check if the clock is already enabled
It is possible for a guest to issue a PM_CLOCK_ENABLE call for a clock
that:
- it is already enabled because it is a critical clock
- it is not related to any of the devices exposed to the guest

For instance, dom0 would try to enable VERSAL_PM_CLK_UART0_REF even if
UART0 is not exposed to dom0.

Instead of returning failure to the guest which would trigger a stack
trace, check if the clock is already enabled. If it is already enabled
simply return success without making any additional calls.

To check whether the clock is already enabled, we use PM_CLOCK_GETSTATE,
which is documented in UG1200 [1] at page 32. The return format is
detailed in UG643 [2] at page 335.

[1] https://www.xilinx.com/support/documentation/user_guides/ug1200-eemi-api.pdf
[2] https://www.xilinx.com/support/documentation/sw_manuals/xilinx2021_1/oslib_rm.pdf

Signed-off-by: Stefano Stabellini <stefano.stabellini@xilinx.com>