xen/eemi: Add call to set SGI interrupt
This calls are rejected from xen for now.
As of now there is no way to notify SGI interrupt
from xen to linux. If SGI registration is allowed
it is possible that we get panic() message.

To avoid this, reject SGI registration calls to TF-A
until we fix way to notify all SGI to linux.

Signed-off-by: Tanmay Shah <tanmay.shah@xilinx.com>
[stefano: fixed code style]
Signed-off-by: Stefano Stabellini <stefano.stabellini@xilinx.com>
Reviewed-by: Stefano Stabellini <stefano.stabellini@xilinx.com>