xen/arm: write boot-time MPU configuration to MPU registers
We have created a boot-time MPU configuration for Xen different
sections and early UART (if configured) in memory. But this
configuration will not take effect until we have written it
to MPU protection region registers and enable MPU.

So in this patch, we will write this boot-time configuration
to MPU hardware. But Xen text and data are in an enabled MPU
protection region, we can't disable this protection region
and update it (disable a protection region where the accessing
data is located will cause unpredictable behavior). So we have
to disable the whole MPU hardware and update the protection
regions. And then re-enable the MPU hardware.

After MPU is disabled, in the worst case, the default background
MPU attributes will be set to device memory. To avoid unexpected
unaligment access fault during MPU disabled, we will use assembly
code to update MPU protection region registers.

And like MMU system force WNX after setting pagetables to MMU
hardware, we also force WNX after setting MPU configuration
to MPU hardware. To avoid to many #ifdef in start_xen, we use
setup_mm_data for MMU and MPU systems to setup boot-time pagetables
and protection regions.

Issue-Id: SCM-4571
Change-Id: I4a98be605b022074dcca445ad5db5efb86f5ac03
Signed-off-by: Wei Chen <wei.chen@arm.com>
Signed-off-by: Penny Zheng <penny.zheng@arm.com>