xen/arm: implement early_fdt_map for MPU systems
For MPU system, device tree binary can be packed in to Xen
image, or provided by bootloader throuhg x0 which is the same
way of MMU system. In both method, before we want to access
device tree binary, we have to map it into an MPU protection
region with proper attributes and permissions. This is the
same requirement as MMU system. But the different is that,
for MMU system, we only require the address of device tree
be 8 bytes alignment. So that we always access the magic and
size fields of the FDT header after mapping the first chunk.
But for MPU system, we need to consider the device tree base
address meet the requirement of MPU region address alignment
(64 bytes).

So in this patch, we consider these two requirements in BUILD
and runtime check before mapping. And in this time, we will
use a new protection region for device tree binary, this means
we will not update a protection region with activities, so we
do not need to disable the whole MPU system. In this case, we
use an unified accessor "access_protection_region" to map
device tree binary to one MPU protection region.

This unified accessor "access_protection_region" can be used
for both reading/writing all possible 256 MPU protection regions.
Armv8-R can support up to 256 MPU protection regions. Each pair
of PRBAR<n>_EL2/PRLAR<n>_EL2 could only provide access to 16
different MPU protection regions, so users also need to set
PRSELR_EL2.REGION to refer to the certain one
(PRSELR_EL2.REGION<7:4>:n).

Futher more, when include device tree binary in dtb.S, we
haven't considered the MPU's alignment as a requirement, so
add a PAGE_SIZE alignment before it to make device tree binary
physical address can meet above two requirements. And PAGE_SIZE
alignment would be a suitable value for MMU and MPU systems.

Change-Id: Icc3ffff6bbcbc7aec62f26c302418785ad53138d
Issue-Id: SCM-4571
Signed-off-by: Wei Chen <wei.chen@arm.com>
Signed-off-by: Penny Zheng <penny.zheng@arm.com>