xen/arm: relocate MPU configuration map from heap
boot_mpumap, as boot-time Xen MPU memory region configuration,
which only supports default MPU protection regions size
(16 as ARM_DEFAULT_MPU_PROTECTION_REGIONS), so it is limited,
and not acceptable in runtime. And if we use the max supported
MPU regions' number to define MPU configuration MPU, it would
be a waste of Xen BSS section, because most of the MPU systems
will not implement 256 MPU protection regions.

Armv8-R provide a register MPUIR_EL2 to indicate the number of
implemented MPU protection regions in EL2. We can read it to
get the accurate number of EL2 MPU protection regions. So after
heap has been created, we can allocate Xen MPU map from heap,
named after xen_mpumap, based on the accurate number.

After that, we will use copy_from_paddr to copy the exisited
MPU configuration from boot_mpumap to xen_mpumap. Since FIXMAP
is not supported on MPU system, so we stub set/clean_fixmap
for copy_from_paddr in MPU systems in this patch.

Also, this update is only applied to MPU systems, so we add
update_mm to wrapper it, and further updates for MPU systems.
And stub update_mm for MMU system to avoid #ifdef.

Change-Id: Iedc4a8d61dc124a4a56108144c113b9fda8a6e95
Issue-ID: SCM-4571
Signed-off-by: Wei Chen <wei.chen@arm.com>
Signed-off-by: Penny Zheng <penny.zheng@arm.com>