xen/arm: implement ioremap_attr for MPU system
As MPU system doesn't have virtual address, it actually does
not have the ability to remap a physical address to a virtual
address with the same or different attributes. But ioremap_*
is used widely in Xen code. It's impossible for us to use #ifdef
MPU to gate everywhere. But it's still not good for us to return
input physical address directly as ioremap_*'s return address.
ioremap_* will require some changes of memory attributes with
new virtual address. If we return a cacheable physical address
for ioremap_wc, this will make things unstable, cache-coherent
issue will be triggered.

So in this patch, we still try to return the input physical
address as remapped new virtual address. But before that we
will check the required attributes of ioremap_*. If current
attributes are the same with required attributes, we can
return it. Otherwise, we will assert and print some messages
to notify users, they have to change the caller function not
to use ioremap to ask a new attributes.

In this way, we can keep most callers of ioremap_nocache for
device memory and ioremap_cache for normal memory unchanged.
Just some ioremap_wc for guest kernel and initrd loader need
to do some adaptions. We adapt them in this patch as well.

Issue-ID: SCM-4571
Change-Id: I206bfc6c72557cec5187012ed6fc26bb3b85d1f2
Signed-off-by: Wei Chen <wei.chen@arm.com>
Signed-off-by: Hongda Deng <Hongda.Deng@arm.com>