xen/arm: map boot module section to MPU protection region
In order to reduce the usage of MPU protection regions, we limit
all boot modules must be located in mpu,boot-module-section. In
this case, we one or minimal MPU protection regions to map boot
modules memory.

In this patch, we map boot module section to MPU protection
regions, and reclaim them in discard_initial_modules.

Issue-ID: SCM-4571
Change-Id: I74e3957a679853ed9760de893bd68f339300de03
Signed-off-by: Wei Chen <wei.chen@arm.com>
Signed-off-by: Penny Zheng <penny.zheng@arm.com>