xen/arm: Use secure hypervisor timer for Armv8-R
As Armv8-R only has secure state, so in secure EL2 we have
to use the secure EL2 hypervisor timer for Xen.

In this patch, we introduce a Kconfig option ARM_SECURE_STATE.
With this new Kconfig option, we can re-define the timer's
system register name in different secure state, but keep the
timer code flow unchanged.

Change-Id: I7ba7f10102cbbfce4af81067746c24ab3b59ad2a
Issue-Id: SCM-4571
Signed-off-by: Wei Chen <wei.chen@arm.com>
Signed-off-by: Henry Wang <Henry.Wang@arm.com>