xen/arm: Disable GIC two security state and Arm SMMU for Armv8R
Armv8R only has one secure state, so we have to check the GIC's
security states, if the GIC support two security state on an Armv8R
platform, we have to disable it.

And there will no be any SMMU on Armv8R platforms, so we disable
Arm SMMU for Armv8 in this patch too.

Change-Id: I7d91342e237fbf496eff1781b3457f0bc687aa09
Signed-off-by: Wei Chen <wei.chen@arm.com>