xen/arm: introduce arch_init_finialize for MPU system
On an MPU system, the number of MPU Protection Regions are
limited, so at the end of Xen boot stage, we have to reclaim
some Protection Regions that are used by boot stage modules.
The reclaim of MPU Protection Regions will cause some holes
in MPU configuration map. So we need to re-order the MPU
configuration map to remove the holes and merge some adjacent
memory regions with the same attributes to one MPU Protection
region. This will save the usage of MPU Protection Regions.

In this patch we introduced arch_init_finialize for MPU system
to such resource reclaim and re-order. And we also defer domain
unpause operaton after arch_init_finialize in this patch.
Otherwise the unpaused domains will run with a temporary,
unreliable EL2 MPU configuration. This will cause unexpected
behaviors for guest.

Change-Id: I86a661f0cab36c452d274e0930f1f096d3366599
Signed-off-by: Wei Chen <wei.chen@arm.com>