xen/arm: implement p2m_init for MPU system
The VMID allocator has been prepared in previous commit, so in this
patch we can start to do some P2M initialization for MPU guests.
On an MPU system, P2M need a table to store guest's stage 2 MPU
configuration. This behavior will be done in p2m_init in this patch.

This patch will allocate VMID and use this VMID to setup VSCTLR_EL2
for guests for MPU system. This is because, On MPU system, we don't
have VTTBR_EL2 to hold the base address of the translation table for
the stage 2 translation and VMID info.

In order to store per-domain register VSCTLR_EL2 value, this commit
also introduces a new field vsctlr in struct p2m_domain, and a helper
generate_vsctlr to generate on runtime.

For some p2m functions that will not be supported on MPU systems, we
simply return -EINVAL/false or 0 to make common code untouched.

As MPU system doesn't have P2M pool, so we implement empty functions
for P2M pool: p2m_get/set_allocation and p2m_teardown_allocation.

Change-Id: I0592db41a06965ae1fb1d8f8be1bec9da3860b64
Issue-ID: SCM-2764
Signed-off-by: Penny Zheng <penny.zheng@arm.com>
Signed-off-by: Wei Chen <wei.chen@arm.com>