xen/arm: implement p2m_get_entry/p2m_set_entry for MPU systems
On MMU system, p2m_set_entry is responsible for
creating/updating/removing an entry(PTE) in page table.

However, since the granularity of P2M memory mapping on
MPU system is not page, but a region as MPU protection
region, which contains multiple pages, this commit tries
to realize similar functionality on MPU system, but the
entry is not a PTE, just a MPU protection region. In order
to implement p2m_set_entry/p2m_get_enry we introduced
following connets in this patch:

p2m_set_permission will be used by p2m_set_entry on MPU
system. So we introduce it in this patch too. Since
mem_access is NOT in use for the domain on MPU system,
here we only intends to check if it is p2m_access_rwx,
the only valid value.

A new function region_to_p2m_entry, which is the equivalence
of mfn_to_p2m_entry in MPU system, and it tries to construct
standard entry in P2M MPU memory mapping table.

This patch also defines a new variable in struct p2m_domain,
nr_regions, to record the number of MPU Protection Regions
in P2M MPU memory mapping table.

In MMU system, page table entry is using 4 bit field to store
p2m type, so in order to achieve the same effect, we are
taking the 4 most significant bit field in prbar to also
store p2m type info. These 4 bits are reserved bits, so they
will be ignored by hardware.

p2m_get_entry is used to get the details of a given gfn, we
reuse the same interfaces in MMU system here, for MPU system.
The core implementation resides on p2m_get_region, and it is
to get the details of a guest memory range, [gfn, gfn + nr_gfns).
If the range is present, only the starting MFN(mfn) will be
returned and also p2m type filled up. Due to GFN == MFN on MPU
system, the whole physical memory range could be easily deduced,
that is, [mfn, mfn + nr_gfns). If the range is not present,
INVALID_MFN will be returned. And for p2m_get_entry in MPU
system, since an entry means a page-sized memory region, the
implementation is to call p2m_get_region with nr_gfns == 1.

Change-Id: I954f37df6a0ada94520fffb2699ff5b826107136
Issue-ID: SCM-2784
Signed-off-by: Penny Zheng <penny.zheng@arm.com>
Signed-off-by: Wei Chen <wei.chen@arm.com>