xen/arm64: move MMU related code from head.S to head_mmu.S
There are lots of MMU specific code in head.S. This code will not
be used in MPU systems. If we use #ifdef to gate them, the code
will become messy and hard to maintain. So we move MMU related
code to head_mmu.S, and keep common code still in head.S.

As we need to access "fail" and "puts" functions out of assembly
file, so we have to export them in this patch. And the assembly
macros: adr_l and load_paddr will be used by MMU and MPU later,
so we move them to macros.h.

Change-Id: Ic095c7bdf0cecad012cc72226c83c86b010b3966
Issue-Id: SCM-4571
Signed-off-by: Henry Wang's avatarHenry Wang <Henry.Wang@arm.com>
Signed-off-by: default avatarWei Chen <wei.chen@arm.com>
Signee-off-by: Ayan Kumar Halder <ayan.kumar.halder@amd.com>