xen/arm: vIOMMU: Add command line option dom0_viommu and its handling
In order to control whether to enable support for virtual IOMMU for
dom0 or not, introduce dom0_viommu boolean command line option.

Take the opportunity to add a check in make_hwdom_viommu_node if
viommu_list head is NULL in which case the vIOMMU node should not be
created.

Signed-off-by: Michal Orzel <michal.orzel@amd.com>