xen: Always set XENFEAT_not_direct_mapped if vIOMMU enabled
At the moment, Linux kernel sets dma operations to Xen SWIOTLB ones for
direct-mapped domains (by checking if a flag XENFEAT_direct_mapped is set)
overriding the IOMMU operations. This prevents the nested IOMMU
configuration to work properly. This patch adds a quirk to always set
XENFEAT_not_direct_mapped in case a domain is running with virtual IOMMU
enabled. In the future, we will need a proper handling in Linux.

Signed-off-by: Michal Orzel <michal.orzel@amd.com>
Reviewed-by: Stefano Stabellini <stefano.stabellini@amd.com>