xen/arm: AArch32_V8R: Change the count of mmu table entries from u64 to u32
set_boot_mpumap() assumes that r0 is the count of mmu table entries
(ie table length) and r1 is the base address of the table.
However, in arm32, a u64 parameter uses two 32 bit registers. So, in
this case, r0 and r1 will represent the table length and r2 will be
the base address of the table.
Therefore, instead of providing a separate implementation of
set_boot_mpumap() for arm32, we just change the prototype to use u32
parameter for table length.

Signed-off-by: Ayan Kumar Halder <ayan.kumar.halder@amd.com>
Acked-by: Stefano Stabellini <stefano.stabellini@amd.com>