xen/arm: AArch32_V8R: Adapt p2m for AArch32_V8R
Defined pr_t for AArch32. This needs to be compatible with AArch64
(ie the size needs to be 128 bits).
Also defined the structures for protection region base address and
protection region limit address hw registers.

So pr_t contains protection region base address register (32 byte),
protection region limit address register (32 byte) and type (64 byte).

In case of AArch64, one uses the unused bits of PRBAR to store the type
of MPU region. This is not possible for AArch32 as there are no unused
bits. So, we store the type as a separate field in pr_t. Also, in the
case of PRLAR, AArch64 has a NS bit which is not present in AArch32.
Thus, we have used appropriate #ifdef to differentiate between the two.

Signed-off-by: Ayan Kumar Halder <ayan.kumar.halder@amd.com>
Acked-by: Stefano Stabellini <stefano.stabellini@amd.com>