xen/arm: AArch32_V8R: Define the functions to access HPRBAR and HPRLAR
Defined a set of macros to configure HPRBAR and HPRLAR.
These are used to create protection regions for the following
during the boot stage :-
1. kernel text
2. kernel read-only data
3. kernel read-write data
4. kernel init text
5. kernel init data
6. kernel BSS
7. Earlyprintk console

Thus, we have defined ARM_DEFAULT_MPU_PROTECTION_REGIONS=16 as the
maximum number of MPU regions to be configured statically for booting
Xen. This defines the size of boot_mpumap[] table.

Also, it is to be noted that HPRBAR and HPRLAR are 32 bit registers on
CONFIG_AARCH32_V8R.
Thus, we have used appropriate macros to different their accesses from
CONFIG_ARM64_V8R.

Signed-off-by: Ayan Kumar Halder <ayan.kumar.halder@amd.com>
Acked-by: Stefano Stabellini <stefano.stabellini@amd.com>