xen/arm: AArch32_V8R: Use 32 bits for the physical address
In the case of MPU (unlike MMU), there is a 1-1 mapping of virtual address
to physical address (ie VA = PA).

However, the physical addresses defined for aarch32 is u64. This is a problem
for aarch32 mpu systems as the physical addresses = virtual addresses = 32 bits.

Thus, when the memory region for FDT is mapped, it returns the virtual address
(which is the same as physical address but truncated upto the lower 32 bits).
Similar logic has been used to convert machine address to virtual address and
for ioremap as well.

We do not support physical addresses beyond 32 bits.
As this logic will fail when FDT physical address is more than 32 bits, we have
added a BUG() to catch this.

Thus, the following functions has been adapted :-
early_fdt_map()
maddr_to_virt()
ioremap_attr()

Also disable "BUILD_BUG_ON(PADDR_BITS >= BITS_PER_LONG)" as PADDR_BITS = 40
and BITS_PER_LONG = 32.

Signed-off-by: Ayan Kumar Halder <ayan.kumar.halder@amd.com>
Signed-off-by: Stefano Stabellini <stefano.stabellini@amd.com>
Acked-by: Stefano Stabellini <stefano.stabellini@amd.com>