xen/pci: Update conditions to handle hwdom using vpci
Add helper macro is_hardware_pci_domain(...) to check if the passed
domain is an hardware domain and it is not using the emulated pci
bus from Xen.

Update conditions to let the hardware domain that is using the vpci
bus emulated by Xen, to use the right addresses instead of the real
host addresses.

When using the emulated bus, update init_bars to don't write in the
bars configuration as it will erase the previous firmware
configuration.

Signed-off-by: Luca Fancellu <luca.fancellu@arm.com>
Signed-off-by: Stewart Hildebrand <stewart.hildebrand@amd.com>
Signed-off-by: Stefano Stabellini <stefano.stabellini@amd.com>
Acked-by: Stefano Stabellini <stefano.stabellini@amd.com>