vsmmuv3: Add dma-coherent property to exposed vSMMUv3 nodes
The latest VFIO framework in Linux (>=6.0) requires both the SMMU and
device we want to use to be dma coherent. This is because from userspace
we cannot (easily) perform cache maintenance operations.

DMA coherency for SMMU means that the page table walk made by SMMU is
cache coherent with CPU. On versal-net SMMU is enabled together with
CCI, so apart from the HW advertising the coherent table walk as
observed by bit IDR0_COHACC of ARM_SMMU_IDR0 register, the CCI will take
care of proper cache coherency.

Signed-off-by: Michal Orzel <michal.orzel@amd.com>
Acked-by: Stefano Stabellini <stefano.stabellini@amd.com>